backend/requirements.txt
fastapi==0.110.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.29
uvicorn==0.29.0
bcrypt==4.1.2
PyJWT==2.8.0

